New Spock unit test covers existing functionality
Before we go about refactoring BestellGenerator it makes sense to cover
the existing behaviour with unit tests. The class is broken enough as
is, I do not want to break it even more but improve it. At least for
more trivial refactorings now there is a series of regression tests.

The tests cover
  - permitted and forbidden command line syntax,
  - behavioural tests (which parameters trigger which calls in
    BgControls with which parsed parameters?),
  - a few TODOs I noticed when writing tests.

What is special about the tests?
  - They use the Spock framework rather thasn plain vanilla JUnit 4.x.
    Consequently they are written in Groovy.
  - @Unroll is used for easy test parametrisation.
  - The JUnit plug-in 'system-rules' by Stefan Birkner is used to catch
    System.exit() calls which would otherwise interrupt test execution.