Be more restrictive with CLI parsing: disallow unknown parameters
Actually I think that the effort to do this does not really pay off, but
maybe somebody wants to see how easily it can be done.

Formerly legal command line test cases with unknown parameters have been
moved to the section with expected failures, i.e. a usage message will
be shown for them now and the test suite checks this.